feat: Routing
The main page contains a link to the search page. When the link is
clicked, the search page is displayed and the URL in the browser’s
address bar is /search.

The search page contains a link to the main page. When the link is 
clicked, the main page is displayed and the URL in the browser’s address
bar is /.